requirements.txt
Pillow==9.3.0
streamlit==1.14.0
torch==1.13.0
torchvision==0.14.0
